+++ lib/ig_markets/streaming/account_state.rb
@@ -125,7 +125,8 @@ module IGMarkets
 
       def convert_market_status(status)
         { auction: :on_auction, auction_no_edit: :on_auction_no_edits, closed: :closed, edit: :edits_only,
-          offline: :offline, suspended: :suspended, tradeable: :tradeable }.fetch status
+          offline: :offline, suspended: :suspended, tradeable: :tradeable, tradeable_no_edit: :tradeable_no_edit
+        }.fetch status
       end
 
       def on_position_update(position_update)